Refactor `did-ion` library.
