+++ Cargo.toml
@@ -77,7 +77,7 @@ members = [
   # "did-sol",
   "did-pkh",
   # "did-onion",
-  # "did-ion",
+  "did-ion",
   # "did-webkey",
   # "vc-test",
   # "did-test",
