+++ did-ion/Cargo.toml
@@ -19,17 +19,17 @@ ed25519 = ["ssi-jws/ed25519"]
 rsa = ["ssi-jws/rsa"]
 
 [dependencies]
-ssi-dids = { path = "../ssi-dids", version = "0.1", features = ["http"] }
-ssi-core = { path = "../ssi-core", version = "0.1" }
-ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false }
-ssi-jws = { path = "../ssi-jws", version = "0.1", default-features = false }
-ssi-jwt = { path = "../ssi-jwt", version = "0.1", default-features = false }
-async-trait = "0.1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_jcs = "0.1"
-anyhow = "1.0"
-thiserror = "1.0"
+ssi-verification-methods.workspace = true
+ssi-dids.workspace = true
+ssi-jwk.workspace = true
+ssi-jws.workspace = true
+ssi-jwt.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_jcs.workspace = true
+iref.workspace = true
+thiserror.workspace = true
+
 base64 = "0.12"
 sha2 = "0.10"
 json-patch = "0.2.6"
